server/requirements.txt
gradio>=4.0.0
requests>=2.28.0
pydantic>=2.0.0
openenv-core[core]>=0.2.1
